Changing type of field responsible for BigInt value

+++ BigInteger/BigInt.cs
@@ -7,6 +7,6 @@ namespace Cryptography
     public sealed partial class BigInt : IComparable<BigInt>
     {
         private bool negative = false;
-        private List<int> numericalRank = new List<int>();                 // big number will be presented as list of integers presenting one digit of number (старшие разряды числа будут дальше идти по списку)       
+        private List<byte> numericalRank = new List<byte>();                 // big number will be presented as list of integers presenting one digit of number (старшие разряды числа будут дальше идти по списку)       
     }
 }
